Read options from express
