+++ examples/README.md
@@ -10,12 +10,12 @@ If you want to use bracket with node and express, it's recommended to use [conso
 
 An example of how to use consolidate can be done running.
 
-- `$ cd examples/node`
-- `$ npm install express consolidate.js`
-- `$ node consolidate.js`
+- `$ cd examples/node/consolidate`
+- `$ npm install express consolidate`
+- `$ node .`
 
 It is also possible to use the library directly. You can see an example by running the example.
 
-- `$ cd examples/node`
+- `$ cd examples/node/simple`
 - `$ npm install express`
-- `$ node simple.js`
+- `$ node .`
