+++ examples/node/consolidate/views/_layout/master.brkt.html
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <title>Consolidate example</title>
+  </head>
+  <body>
+    [[# body() ]]
+  </body>
+</html>
