+++ examples/node/consolidate/views/home.brkt.html
@@ -1,0 +1,11 @@
+---
+master: _layout/master.brkt.html
+---
+
+[[## body()
+  Hello from home.brkt.html <br />
+  Model data:
+  <pre>
+    [[= model.test1 ]]
+  </pre>
+#]]
