+++ examples/node/simple/index.js
